chore(gesture-feedback): satisfy detekt and dump the new public API
CI was red on two checks:

- `apiCheck` — the new gesture-feedback surface was never dumped. The diff is
  purely additive (122 insertions, 0 deletions), so binary compatibility holds.
- detekt — four violations, all in the new code:
  - `MatchingDeclarationName` on two files, fixed the way this codebase already
    fixes it (SurfaceMirroring.kt, Splat.kt): split into a same-package file
    named after its single top-level class, so imports are unaffected.
  - `TooManyFunctions` on `NodeGestureDelegate` (26 > 25) — the two press/release
    notifiers move out as internal extensions in `NodeEditingDispatch.kt`.
  - `ComplexCondition` on the four-way null check that fits the base ellipse —
    the two projected chords are now paired up front, which also names what the
    check is actually asserting.

Verified locally: detekt, apiCheck, lintDebug, and compilation of the library
and the demo app all pass.
